Add a UserContext with a hardcoded user value.
I would like some information about the user available throughout the
app. There's no redux, so I'm setting up a context.

We also need the frontend to check with the backend to see if we're
logged in. If I restart the backend, and F5 the frontend.. since the dev
sessions are not persisted, I should be logged out.  I am actually
logged out, but the frontend doesn't realize it.

So, the next step is the UserProvider will check with the backend to see
who we're logged in as.  If we have auth tokens set, but the backend
says we're not logged in, then we can automatically clear those tokens.