Add Profile endpoint and UserContext
- UserContext should automatically be populated with info about the user
  (which info is customizable, look in user.ex)
- Frontend will automatically get rid of auth/renew tokens that no
  longer work, you should no longer appear to be logged in while you're
  logged out

Wrestled with hooks some here and had a fair number of infinite loop
problems.  Not sure if my "useAuthDataApi" is the right abstraction..
it's getting tough to work with.

Also, I tried implementing an "options object parameter" for
useAuthDataApi, but it always lead to infinite loops, since it's not
memoized.  Not sure how to do it with default options merging in a
custom hook.

Maybe some of this stuff would have been easier in redux?